Fix interceptors
