+++ data-runtime/src/main/java/io/micronaut/data/runtime/intercept/DefaultFindPageInterceptor.java
@@ -15,22 +15,9 @@
  */
 package io.micronaut.data.runtime.intercept;
 
-import io.micronaut.aop.MethodInvocationContext;
 import io.micronaut.core.annotation.NonNull;
-import io.micronaut.core.util.CollectionUtils;
-import io.micronaut.data.annotation.Query;
 import io.micronaut.data.intercept.FindPageInterceptor;
-import io.micronaut.data.intercept.RepositoryMethodKey;
-import io.micronaut.data.model.CursoredPage;
-import io.micronaut.data.model.Page;
-import io.micronaut.data.model.Pageable;
-import io.micronaut.data.model.Pageable.Cursor;
-import io.micronaut.data.model.Pageable.Mode;
-import io.micronaut.data.model.runtime.PreparedQuery;
 import io.micronaut.data.operations.RepositoryOperations;
-import io.micronaut.data.runtime.operations.internal.sql.DefaultSqlPreparedQuery;
-
-import java.util.List;
 
 /**
  * Default implementation of {@link FindPageInterceptor}.
@@ -40,7 +27,7 @@ import java.util.List;
  * @author graemerocher
  * @since 1.0.0
  */
-public class DefaultFindPageInterceptor<T, R> extends AbstractQueryInterceptor<T, R> implements FindPageInterceptor<T, R> {
+public class DefaultFindPageInterceptor<T, R> extends DefaultAbstractFindPageInterceptor<T, R> implements FindPageInterceptor<T, R> {
 
     /**
      * Default constructor.
@@ -48,48 +35,5 @@ public class DefaultFindPageInterceptor<T, R> extends AbstractQueryInterceptor<T
      */
     protected DefaultFindPageInterceptor(@NonNull RepositoryOperations datastore) {
         super(datastore);
-    }
-
-    @Override
-    public R intercept(RepositoryMethodKey methodKey, MethodInvocationContext<T, R> context) {
-        Class<R> returnType = context.getReturnType().getType();
-        if (context.hasAnnotation(Query.class)) {
-            PreparedQuery<?, ?> preparedQuery = prepareQuery(methodKey, context);
-
-            Iterable<?> iterable = operations.findAll(preparedQuery);
-            List<R> results = (List<R>) CollectionUtils.iterableToList(iterable);
-            Pageable pageable = getPageable(context);
-            Long totalCount = null;
-            if (pageable.requestTotal()) {
-                PreparedQuery<?, Number> countQuery = prepareCountQuery(methodKey, context);
-                Number n = operations.findOne(countQuery);
-                totalCount = n != null ? n.longValue() : null;
-            }
-
-            Page<R> page;
-            if (pageable.getMode() == Mode.OFFSET) {
-                page = Page.of(results, pageable, totalCount);
-            } else if (preparedQuery instanceof DefaultSqlPreparedQuery<?, ?> sqlPreparedQuery) {
-                List<Cursor> cursors = sqlPreparedQuery.createCursors((List<Object>) results, pageable);
-                page = CursoredPage.of(results, pageable, cursors, totalCount);
-            } else {
-                throw new UnsupportedOperationException("Only offset pageable mode is supported by this query implementation");
-            }
-            if (returnType.isInstance(page)) {
-                return (R) page;
-            } else {
-                return operations.getConversionService().convert(page, returnType)
-                        .orElseThrow(() -> new IllegalStateException("Unsupported page interface type " + returnType));
-            }
-        } else {
-
-            Page page = operations.findPage(getPagedQuery(context));
-            if (returnType.isInstance(page)) {
-                return (R) page;
-            } else {
-                return operations.getConversionService().convert(page, returnType)
-                        .orElseThrow(() -> new IllegalStateException("Unsupported page interface type " + returnType));
-            }
-        }
     }
 }
